changed selfhosted package name back to apiserver and moved 2 files to deserialization package
