+++ src/main/java/com/laserfiche/api/client/apiserver/TokenClient.java
@@ -1,4 +1,4 @@
-package com.laserfiche.api.client.selfhosted;
+package com.laserfiche.api.client.apiserver;
 
 import com.laserfiche.api.client.model.CreateConnectionRequest;
 import com.laserfiche.api.client.model.SessionKeyInfo;
