+++ src/main/java/com/laserfiche/api/client/httphandlers/UsernamePasswordHandler.java
@@ -2,8 +2,8 @@ package com.laserfiche.api.client.httphandlers;
 
 import com.laserfiche.api.client.model.CreateConnectionRequest;
 import com.laserfiche.api.client.model.SessionKeyInfo;
-import com.laserfiche.api.client.selfhosted.TokenClient;
-import com.laserfiche.api.client.selfhosted.TokenClientImpl;
+import com.laserfiche.api.client.apiserver.TokenClient;
+import com.laserfiche.api.client.apiserver.TokenClientImpl;
 
 import java.util.concurrent.CompletableFuture;
